적 상태별 행동 다르게 하기 위해 EnemyState 상태 클래스 생성 및 적용

+++ UnitTest/Character/Agent.cpp
@@ -122,7 +122,7 @@ void Agent::Update()
     float aimAngle = D3DXVec3Dot(&Values::UpVec, &agentLookAtAxis);
 
     aimAngle = std::acos(aimAngle);
-    aimAngle = D3DXToDegree(aimAngle);
+    aimAngle = (float)D3DXToDegree(aimAngle);
 
     switch (combatState)
     {
